# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: TabMail API
  version: "2.0"
  description: |
    TabMail is a multi-tenant, API-first disposable email service with both
    inbound (SMTP receive) and outbound (SMTP send) capabilities.

    ## Quick Start for Bot/Automation Users

    ### Step 1: Get an API Key
    ```bash
    # Login to get a JWT token
    curl -X POST https://your-tabmail/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email":"you@example.com","password":"your-password"}'

    # Create an API Key with the JWT
    curl -X POST https://your-tabmail/api/v1/keys \
      -H "Authorization: Bearer <jwt_token>" \
      -H "Content-Type: application/json" \
      -d '{"label":"my-bot","scopes":["mailboxes:write","messages:read","send:write"]}'
    ```

    ### Step 2: Create a Disposable Mailbox
    ```bash
    # Get a suggested address
    curl https://your-tabmail/api/v1/domains/{domain_id}/suggest-address \
      -H "X-API-Key: tabmail_..."

    # Create the mailbox
    curl -X POST https://your-tabmail/api/v1/mailboxes \
      -H "X-API-Key: tabmail_..." \
      -H "Content-Type: application/json" \
      -d '{"full_address":"suggested-addr@domain.com","zone_id":"..."}'
    ```

    ### Step 3: Receive Emails (Poll or SSE)
    ```bash
    # Poll for messages
    curl https://your-tabmail/api/v1/mailbox/addr@domain.com \
      -H "X-API-Key: tabmail_..."

    # Or use Server-Sent Events for real-time
...
      -H "X-API-Key: tabmail_..."
    ```

...

...

...

...

...

...

...

...
servers:
...
security:
  - {}
  - APIKey: []
components:
  securitySchemes:
    APIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key
    MailboxToken:
      type: http
      scheme: bearer
      bearerFormat: TabMail-Mailbox-Token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key (alias for APIKey)
  schemas:
...
    Meta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
...
    DomainRoute:
      type: object
      properties:
        id: { type: string, format: uuid }
        zone_id: { type: string, format: uuid }
        route_type: { type: string, enum: [exact, wildcard, deep_wildcard, sequence] }
        match_value: { type: string }
        range_start: { type: integer, nullable: true }
        range_end: { type: integer, nullable: true }
        auto_create_mailbox: { type: boolean }
        retention_hours_override: { type: integer, nullable: true }
        access_mode_default: { type: string, enum: [public, token, api_key] }
        created_at: { type: string, format: date-time }
    SuggestedAddress:
      type: object
      properties:
        zone_id: { type: string, format: uuid }
        base_domain: { type: string }
        domain: { type: string }
        subdomain_label: { type: string, nullable: true }
        local_part: { type: string }
        address: { type: string }
        mode: { type: string, enum: [mailbox, subdomain] }
        algorithm: { type: string }
    Mailbox:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        zone_id: { type: string, format: uuid }
        local_part: { type: string }
        resolved_domain: { type: string }
        full_address: { type: string }
        access_mode: { type: string, enum: [public, token, api_key] }
        retention_hours_override: { type: integer, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 30
        minimum: 1
        maximum: 100
      description: Items per page
paths:
...
  /api/v1/resources/domains/{id}/suggest-address:
    get:
      tags: [Domains]
      summary: Generate a randomized address for an open domain resource
      operationId: suggestOpenAddress
      security: [ {}, { BearerAuth: [] } ]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: subdomain
          in: query
          required: false
          schema: { type: boolean }
          description: Requires allow_random_subdomains on the open resource
      responses:
        "200":
          description: Suggested randomized mailbox address
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/SuggestedAddress" }
        "403": { description: Resource is not open to this viewer }
        "404": { description: Not found }

  # ── Domains / Zones ─────────────────────────────────
...
  /api/v1/domains/{id}/verify:
    post:
      tags: [Domains]
      summary: Trigger DNS verification
      operationId: triggerVerify
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Verification info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      domain: { type: string }
                      txt_record: { type: string }
                      is_verified: { type: boolean }
                      mx_verified: { type: boolean }
                      checks: { type: object }
                      hint: { type: string }
  /api/v1/domains/{id}/verification-status:
    get:
      tags: [Domains]
      summary: Get DNS verification status
      operationId: verificationStatus
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Verification details (TXT, MX, SPF, DKIM, DMARC)
  /api/v1/domains/{id}/suggest-address:
    get:
      tags: [Domains]
      summary: Generate a randomized mailbox address suggestion for a domain zone
      operationId: suggestAddress
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: subdomain
          in: query
          required: false
          schema: { type: boolean }
          description: When true, also randomize a child subdomain under the parent zone
      responses:
        "200":
          description: Suggested randomized mailbox address
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/SuggestedAddress" }
        "400": { description: Bad request }
        "403": { description: Forbidden }
        "404": { description: Not found }

  # ── Domain Routes ───────────────────────────────────
  /api/v1/domains/{id}/routes:
    get:
      tags: [Domain Routes]
      summary: List routes for a zone
      operationId: listRoutes
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DomainRoute" }
    post:
      tags: [Domain Routes]
      summary: Create a routing rule
      operationId: createRoute
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [route_type, match_value]
              properties:
                route_type: { type: string, enum: [exact, wildcard, deep_wildcard, sequence] }
                match_value: { type: string, example: "*.mail.example.com" }
                range_start: { type: integer }
                range_end: { type: integer }
                auto_create_mailbox: { type: boolean, default: true }
                retention_hours_override: { type: integer }
                access_mode_default: { type: string, enum: [public, token, api_key], default: public }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/DomainRoute" }
  /api/v1/domains/{id}/routes/{routeId}:
    delete:
      tags: [Domain Routes]
      summary: Delete a route
      operationId: deleteRoute
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: routeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }

  # ── Mailboxes ───────────────────────────────────────
  /api/v1/mailboxes:
    get:
      tags: [Mailboxes]
      summary: List mailboxes
      operationId: listMailboxes
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Mailbox" }
                  meta: { $ref: "#/components/schemas/Meta" }
    post:
      tags: [Mailboxes]
      summary: Create a mailbox
      operationId: createMailbox
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address: { type: string, example: "user@mail.example.com" }
                password: { type: string }
                access_mode: { type: string, enum: [public, token, api_key], default: public }
                retention_hours_override: { type: integer, nullable: true, example: 24 }
                expires_at: { type: string, format: date-time, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Mailbox" }
        "409": { description: Address already exists }
  /api/v1/mailboxes/{id}:
    delete:
      tags: [Mailboxes]
      summary: Delete a mailbox
      operationId: deleteMailbox
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }

  # ── Messages ────────────────────────────────────────
...